pyproject.toml
[project]
name = "readwise-weekly-digest"
version = "1.0.0"
description = "Automated weekly reading digest generator from Readwise to GitHub"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "functions-framework>=3.4.0",
    "requests>=2.31.0",
    "PyGithub>=1.59.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

# [project.urls]
# Homepage = "https://github.com/yourusername/readwise-weekly-digest"
# Repository = "https://github.com/yourusername/readwise-weekly-digest"

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]